Move layouting service. We see how layouting is only used on msg, we can remove this probably via signal
